major refactoring and useReceiverFormStore
